Prevent kernel panic in uncore on 32 bit with x2APIC.
The uncore code is heavily dependent on the CPU topology mapping code,
so if we do not even have a working topology for the boot CPU we just
give up.

This happens for example on HPE DL360 Gen9 servers with x2APIC enabled
in the firmware settings, where the kernel only finds a single core.